requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.14.2
boto3==1.40.45
botocore==1.40.45
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dill==0.4.0
docling==2.55.1
docling-core==2.48.4
docling-ibm-models==3.9.1
docling-parse==4.5.0
easyocr==1.7.2
et_xmlfile==2.0.0
Faker==37.8.0
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==1.0.1
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.5
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.2
psutil==7.1.0
pyclipper==1.3.0.post6
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rtree==1.4.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.9.30
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
triton==3.4.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.9
